% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stats.R
\name{check_cochran}
\alias{check_cochran}
\title{check Cochran rule}
\usage{
check_cochran(
  efftheo,
  seuil = 80,
  effcritique = 1,
  effcochran = 5,
  verbose = TRUE
)
}
\arguments{
\item{efftheo}{A table of theoretical counts}

\item{seuil}{A numeric between 0 and 100. Default 80\%.
Percentage of cell supposed get a count greater than effcochran (Eg. 80\% should have count above 5.)}

\item{effcritique}{A integer. Default 1. Integer of the minimal count for a given cell}

\item{effcochran}{A integer. Default 5. Integer if the minimal count for maximum seuil in \% of cells}

\item{verbose}{A logical, Default TRUE. Print the conclusion message.}
}
\value{
A logical. TRUE means the Cochran rule is respected, otherwise FALSE.
}
\description{
Check theoretical sample size in table
}
\examples{
\dontrun{
effobs <- table(modified_state[, c("state.region", "election")])
# tableau des effectifs theoriques
efftheo <- as.table(round(rowSums(effobs) \%*\% t(colSums(effobs)) / sum(effobs), 2))
names(dimnames(efftheo)) <- names(dimnames(effobs))
rownames(efftheo) <- rownames(effobs)
colnames(efftheo) <- colnames(effobs)
check_cochran(
  efftheo = efftheo
)
}
}
