% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stats.R
\name{check_fisher}
\alias{check_fisher}
\title{check Fisher good application}
\usage{
check_fisher(tc)
}
\arguments{
\item{tc}{A table of contingency count.}
}
\value{
A logical. TRUE means the fisher "rule" is respected, otherwise FALSE.
}
\description{
In the situation where one variable (Var1) has missing data for one of the levels of another variable (Var2),
this leads to zeros in some cell of the contingency table.
}
\details{
Structural zeros: If these zeros are due to a structural impossibility
(that is, they are expected by the nature of the data, for example,
if a Var1 category does not exist for a Var2 category),
the exact Fisher test may not be appropriate because it assumes that all combinations are possible.

Zeros observed: If these zeros are not structural but simply observed
(that is, they result from sampling and not from a theoretical constraint),
Fisher’s exact test can be used.
However, the results should be interpreted with caution.
The test is still mathematically correct, but zeros may indicate a problem with the data
(for example, a lack of information or sampling bias).

In the 2 cases (structural zeros or observed zeros) we will prefer not to display the fisher statistic here and
indicates that the correct application of the fisher test is not "valid".
}
