% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_normality}
\alias{check_normality}
\title{check_normality}
\usage{
check_normality(x, return_messages = FALSE)
}
\arguments{
\item{x}{Numeric vector to test}

\item{return_messages}{Logical, Should error messages be returned?}
}
\value{
Logical (normality) or
 list with is_normal the logical and its message
}
\description{
Testing for normality with error handling
}
