% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tables.R
\name{compute_correlation_table}
\alias{compute_correlation_table}
\title{compute correlation table}
\usage{
compute_correlation_table(
  dataframe,
  vars = setdiff(colnames(dataframe), varstrat),
  varstrat = varstrat,
  method_corr = "detect_auto",
  digits = 2,
  signif_digits = 4
)
}
\arguments{
\item{dataframe}{A data.frame, tibble or data.table will be converted into data.table.}

\item{vars}{A vector of characters. Names of dataframe's continuous columns to describe.}

\item{varstrat}{A character. Always needed. Name of the continuous var to compute the correlation with.}

\item{method_corr}{A character. Default "detect_auto" meaning the linear link will be detected if the R2 > 0.5.
Otherwise, method to compute correlation is wanted "pearson", "spearman", "kendall".}

\item{digits}{A integer, Default 2. Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer, Default 4. Integer indicating the number of decimal places (signif) for pvalues.}
}
\value{
A final table of stat and correlation.
}
\description{
provide descriptive statistics table for continuous data and its correlation with a continuous varstrat
}
\examples{
\dontrun{
compute_correlation_table(
  dataframe = modified_state,
  vars = c(
    "Population", "Income", "Illiteracy", "Life Exp",
    "HS Grad", "Frost"
  ),
  varstrat = "Area",
  method_corr = "detect_auto",
  digits = 2
)
}
}
