% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_date_tables.R
\name{compute_date_table}
\alias{compute_date_table}
\title{compute_date_table}
\usage{
compute_date_table(
  dataframe,
  vars = setdiff(colnames(dataframe), varstrat),
  varstrat = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table will be converted into data.frame.}

\item{vars}{A vector of characters. Names of dataframe's date columns to describe.
only consider "Date", "POSIXct" or "POSIXt" format.}

\item{varstrat}{A character. Default NULL. Name of the stratification variable,
making groups to compare.}

\item{verbose}{A logical, Default TRUE. Show message. 
Do you want to work in silence? Turn it FALSE.}
}
\value{
A list of tables with nb measures, mean, med, min and max of dates.
}
\description{
compute date table : provide descriptive statistics table for dates
}
\examples{
\dontrun{
dt <- data.frame(
 date1 = as.Date(c("2025-07-30", "2025-07-29", "2025-07-28", "2025-07-30", "2025-07-29", "2025-07-28")), 
 date2 = as.POSIXct(c("2024-06-20", "2024-06-10", "2024-06-01", "2025-07-30", "2025-07-29", "2025-07-28")), 
 fact = as.factor(c(rep("A", 3), rep("B", 3)))
)
get_dates(dataframe = dt, vars = names(dt))
compute_date_table(
  dataframe = dt,
  vars = names(dt)
)
compute_date_table(
  dataframe = dt,
  vars = names(dt), 
  varstrat = "fact"
)
}

}
