% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{compute_frequency_table}
\alias{compute_frequency_table}
\title{compute_frequency_table}
\usage{
compute_frequency_table(
  dt,
  variable_interest,
  varstrat,
  varstrat_levels,
  digits = 1
)
}
\arguments{
\item{dt}{A data.frame}

\item{variable_interest}{A character}

\item{varstrat}{A character}

\item{varstrat_levels}{A character}

\item{digits}{An integer}
}
\value{
table of frequencies
}
\description{
Calculate the frequency table using counts and percentages
}
\keyword{internal}
