% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_precision_digits}
\alias{compute_precision_digits}
\title{compute_precision_digits}
\usage{
compute_precision_digits(
  x,
  stat_type = c("central", "sd"),
  base_decimals,
  max_decimals = 3
)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric
}
\description{
Function to calculate digits according to precision
}
\details{
Because we have established a rule within our quality approach to harmonise 
the accuracy of the figures provided in our reports. 
The rule states that: 
The mean and median will be rounded to one decimal place higher than the 
accuracy of the original value. 
The standard deviation will be rounded to 2 decimal places above the 
precision of the original value, up to a maximum of 3 decimal places.
}
\examples{
\dontrun{
compute_precision_digits(
  c(3.5,3.6,1.5,9), stat_type = "central", base_decimals = 1
)
compute_precision_digits(
  c(3.5526,3.6458,1.512,9.2), stat_type = "central", base_decimals = 4
)
compute_precision_digits(
  c(3.5526,3.6458,1.512,9.2), stat_type = "sd", base_decimals = 4
)
}
}
