% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tables.R
\name{compute_single_factor_stats}
\alias{compute_single_factor_stats}
\title{compute_single_factor_stats}
\usage{
compute_single_factor_stats(
  dataframe,
  var_name,
  varstrat = NULL,
  varstrat_levels = NULL,
  prop_table_margin = 2,
  precision = 1,
  force_generate_1_when_0 = FALSE
)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table will be converted into data.frame.}

\item{varstrat}{A character. Default NULL. Name of the stratification variable,
making groups to compare.}

\item{varstrat_levels}{--here}

\item{prop_table_margin}{A vector giving the margins to split by. Default 2. 
1 indicates rows, 2 indicates columns,
c(1, 2) indicates rows and columns. When x has named dimnames, 
it can be a character vector selecting dimension names.}

\item{precision}{Precision mode: "auto" (adaptive) or numeric (fixed)}

\item{force_generate_1_when_0}{A logical, Default FALSE. If TRUE, 
will test if the unique modality is 0 or "non" and
add the level 1 or "oui" so it can be display in counts. 
Can be combined with simplify to only show the modality (1).}

\item{vars}{A vector of characters. Names of dataframe's factorial columns to describe.}

\item{simplify}{A logical. Default TRUE. Reduce the yes/no  (or 1/0) modalities 
to display only the yes (1).}
}
\value{
a table of n and p for each levels
 + Nb measures and N of missing values (Valeurs_manquantes)
}
\description{
Unit function for a factorial variable
}
\examples{
\dontrun{
data <- data.frame(
  gender = as.factor(c("F", "M", "F", "F", "M")),
  ttt = as.factor(c("a", "a", "a", "b", NA)),
  success = as.factor(c(1, 1, 1, NA, NA))
)
compute_single_factor_stats(dataframe = data, var_name = "gender")
compute_single_factor_stats(dataframe = data, var_name = "success")
compute_single_factor_stats(
  dataframe = data, var_name = "ttt", 
  varstrat = "gender" , varstrat_levels = levels(data$gender)
)
}
}
