% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SMD_tables.R
\name{compute_SMD_table}
\alias{compute_SMD_table}
\title{compute SMD}
\usage{
compute_SMD_table(dataframe, vars, varstrat, precision = 2)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table}

\item{vars}{A vector of characters. Names of dataframe's columns.}

\item{varstrat}{A character. Default NULL. Name of the stratification variable, making groups to compare.}

\item{precision}{Precision mode: "auto" (adaptive) or numeric (fixed)
Integer indicating the number of decimal places (round).}
}
\value{
A table of SMD values for each vars.
}
\description{
Standardized mean difference
}
\details{
Interpretation of SMD values :  
The generally accepted thresholds for interpreting effect size are :  
|SMD| < 0.1: negligible difference,  
0.1 ≤ |SMD| < 0.2: "Imbalance was usually defined as "stddiff" greater than 0.1 or 0.2 (which means the small effect size)."  
0.2 ≤ |SMD| < 0.5: between small and moderate difference,    
0.5 ≤ |SMD| < 0.8: between moderate and large difference,  
|SMD| ≥ 0.8: large difference.  

These thresholds help to assess the balance between groups,  
 especially after matching or adjustment procedures.  

Computation follows {stddiff} r pkg implementation 
(https://cran.r-project.org/web/packages/stddiff/index.html)  

Reading of thresholds are based on :  
  - Cohen, J. (1988).  Cohen J. Statistical Power Analysis for the Behavioral Sciences (2nd ed). Lawrence Erlbaum Associates Publishers: Hillsdale, NJ.   
  - also quoted in https://www.utstat.toronto.edu/brunner/oldclass/378f16/readings/CohenPower.pdf (page 40 – num 57/579)  
  - also quoted in SAS documentation : https://support.sas.com/resources/papers/proceedings12/335-2012.pdf  

Why SMD ?  
In the context clinical trials, the use of statistical tests to assess
baseline comparability is illogical, as noted by Altman in 1985:  
"… performing a significance test to compare baseline variables is to assess 
the probability of something having occurred by chance when we know that it did
occur by chance. Such a procedure is clearly absurd."  
Statisticians and clinical research methodologists have maintained that baseline 
significance testing should not be performed, with Austin noting in 2010:  
"With few exceptions, the statistical literature is uniform in its agreement 
on the inappropriateness of using hypothesis testing to compare the 
distribution of baseline covariates between treated and untreated 
subjects in randomised controlled trials." 
(https://www.sciencedirect.com/science/article/pii/S1836955315000806?via%3Dihub)
}
\examples{
\dontrun{
compute_SMD_table(
  dataframe = modified_state,
  vars = c(
    "Population", "Income", "Illiteracy", "Life Exp", "Murder",
    "HS Grad", "Frost", "Area", "yes_no_french_question", "binary_test",
    "state.division", "state.region"
  ),
  varstrat = "election",
  precision = 2
 )
}
}
