% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_file_production.R
\name{date_sheet}
\alias{date_sheet}
\title{date_sheet}
\usage{
date_sheet(dataframe, vars_dates, varstrat, verbose)
}
\arguments{
\item{dataframe}{A data.frame. Columns must be well formatted with factor or 
numeric class (important). will be converted as data.table inside functions.}

\item{vars_dates}{A vector of characters. Names of dataframe's columns to describe.
Only consider date (subset of vars).}

\item{varstrat}{A character. Name of the stratification variable,
making groups to compare.
Only one varstrat is expected in general.
About 2 crossed varstrats : you can provide "var1*var2" to get a description 
of the 2nd varstrat for each level of the 1st varstrat.}

\item{verbose}{A logical, Default TRUE. Show message. 
Do you want to work in silence? Turn it FALSE.}
}
\value{
list with formatted dates variables description
}
\description{
Compute dates variables sheet for Excel output
}
\details{
Internal function that processes dates variables and prepares
the formatted table for Excel export
}
