% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_decimal_places}
\alias{detect_decimal_places}
\title{detect_decimal_places}
\usage{
detect_decimal_places(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a numeric
}
\description{
Function to detect the number of decimal places in a vector
}
\examples{
\dontrun{
detect_decimal_places(
  c(3.5,3.6,1.5,9)
) # 1
detect_decimal_places(
  c(3.5526,3.6458,1.512,9.2)
) # 4
}
}
