% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect extreme values in a numeric data.table}
\usage{
detect_outliers(
  dt,
  vars,
  method = c("Tukey", "Zscore"),
  threshold = NULL,
  summary = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dt}{A data.frame or data.table with numeric columns only}

\item{vars}{Character vector of variable names to check.
Default: all numeric columns.}

\item{method}{Method to detect outliers: "Tukey" or "Zscore"}

\item{threshold}{Multiplier for the bounds.
Default: 1.5 for Tukey (classic boxplot), 3 for Zscore (3 SD).}

\item{summary}{Logical. If TRUE, returns a summary table (one row per
variable). If FALSE, returns all outlier rows. Default: FALSE.}

\item{verbose}{A logical, Default TRUE. Show message. 
Do you want to work in silence? Turn it FALSE.}
}
\value{
A data.table with outlier details or summary, NULL if none found.
  Check the "Row_index" line to identify the row in the original data set (dt)
   with an extreme value.
}
\description{
Tukey Method will use [Q1 - threshold * IQR  ;  Q3 + threshold * IQR]
Zscore Method will use [mean - threshold * SD  ;  mean + threshold * SD]
}
\details{
You can use either classic threshold 
  (1.5 for Tukey (classic boxplot), 3 for Zscore (3 SD))
  nor apply your own threshold.
}
\examples{
\dontrun{
dt <- data.table::data.table(
  PAM     = c(rnorm(43, 85, 10), 145, 150),
  LACTATE = c(rnorm(44, 1.8, 0.5), 8.2, NA),
  FC      = c(rnorm(44, 75, 10), 180, NA)
)

# Tukey method (default threshold 1.5)
detect_outliers(dt, vars = names(dt), method = "Tukey")

# Zscore method (default threshold 3)
detect_outliers(dt, vars = names(dt), method = "Zscore")

# Summary only
detect_outliers(dt, vars = names(dt), method = "Tukey", summary = TRUE)

# Tukey avec threshold custom (boxplot étendu : 3 au lieu de 1.5)
detect_outliers(dt, vars = names(dt), method = "Tukey", threshold = 3)

}
}
