% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stats.R
\name{estimate_diff_mean}
\alias{estimate_diff_mean}
\title{estimate difference of means}
\usage{
estimate_diff_mean(
  dataframe,
  vars,
  varstrat,
  digits = 1,
  force_non_parametric_test = FALSE
)
}
\arguments{
\item{dataframe}{A data.frame containing data.}

\item{vars}{A vector of characters. Names of dataframe's factorial columns to describe.}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare.}

\item{digits}{A integer, Default 1. Integer indicating the number of decimal places (round).}

\item{force_non_parametric_test}{A logical. Default FALSE. You can turn it TRUE if you want to force the use of
non parametric test, whatever shapiro test said about normality.}
}
\value{
A data.frame,
  in rows the vars and in column the stat used, the difference values and the IC95%
}
\description{
Compute difference of means and its IC between groups
}
\examples{
\dontrun{
estimate_diff_mean(
  dataframe = modified_state[, c("Population", "Income", "election")],
  vars = c("Population", "Income"),
  varstrat = "election"
)
}
}
