% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_cell_content}
\alias{format_cell_content}
\title{format_cell_content}
\usage{
format_cell_content(stats, metric = "mean")
}
\arguments{
\item{stats}{A data.frame}

\item{metric}{A character}
}
\value{
cell formated
}
\description{
format_cell_content
}
\keyword{internal}
