% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_effectif_table}
\alias{format_effectif_table}
\title{Créer le tableau d'effectifs formaté}
\usage{
format_effectif_table(
  sumup,
  variable_interest,
  varstrat,
  varstrat_levels,
  keep_missing_line
)
}
\description{
Créer le tableau d'effectifs formaté
}
