% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_global_summary}
\alias{format_global_summary}
\title{format_global_summary}
\usage{
format_global_summary(
  dataframe,
  variable,
  metric = "mean",
  patient_id = NULL,
  digits_central,
  digits_sd
)
}
\arguments{
\item{dataframe}{A data.frame}

\item{variable}{A character}

\item{metric}{A character}

\item{patient_id}{A character or NULL}

\item{digits_central}{A integer}

\item{digits_sd}{A integer}
}
\value{
global summary formated
}
\description{
format_global_summary
}
\keyword{internal}
