% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_characters}
\alias{get_characters}
\title{get characters}
\usage{
get_characters(dataframe, vars = colnames(dataframe))
}
\arguments{
\item{dataframe}{A data.frame.}

\item{vars}{A vector of characters. Names of dataframe to test.}
}
\value{
A vector of characters : the names of character columns
}
\description{
provide the names of numeric columns
}
\examples{
\dontrun{
get_numerics(
  dataframe = modified_state,
  vars = c("state.division", "state.region", "Income", "Illiteracy")
)
}
}
