% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factors}
\alias{get_factors}
\title{get factors}
\usage{
get_factors(dataframe, vars = colnames(dataframe))
}
\arguments{
\item{dataframe}{A data.frame.}

\item{vars}{A vector of characters. Names of dataframe to test.}
}
\value{
A vector of characters : the names of factorial columns
}
\description{
provide the names of factorial columns
}
\examples{
\dontrun{
}
}
