% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_OR_tables.R
\name{get_OR_univar}
\alias{get_OR_univar}
\title{get OR univariate model}
\usage{
get_OR_univar(
  dataframe,
  dependent_var = "Y",
  explanatory_vars = c("X1", "X2"),
  check_n_levels = FALSE,
  signif_digits = 4
)
}
\arguments{
\item{dataframe}{A data.frame, tibble or data.table.}

\item{dependent_var}{A character. Name of the stratification variable, making groups to compare.}

\item{explanatory_vars}{A vector of characters. Names of dataframe's columns to describe.}

\item{check_n_levels}{A logical, Default FALSE. Maybe wanted on qualitative explanatory_vars, so those with 0 levels are removed.
keep numeric variables too.}

\item{signif_digits}{A integer, Default 4. Integer indicating the number of decimal places (signif) for pvalues.}
}
\value{
A final table with Variable names, Modalites asso to the given OR (IC, p).
}
\description{
provide statistics table of OR, IC and p-values for glm binomial model applied on each explanatory_vars
modele : dependent_var ~ explanatory_vars_i
}
\examples{
\dontrun{
dependent_var <- "binary_test"
explanatory_vars <- c(
  "Population", "Income", "Illiteracy", "Life Exp", "Murder",
  "HS Grad", "Frost", "state.division",
  "var_conti_trap",
  "state.region",
  "var_quali_trap",
  "yes_no_french_question", "all_count_zero",
  "var_ab_NAN"
)
get_OR_univar(
  dataframe = modified_state,
  dependent_var = dependent_var,
  explanatory_vars = explanatory_vars
)
}
}
