% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{handle_force_generate}
\alias{handle_force_generate}
\title{handle_force_generate}
\usage{
handle_force_generate(dt, variable_interest, force_generate_1_when_0)
}
\arguments{
\item{dt}{A data.frame}

\item{variable_interest}{A character}

\item{force_generate_1_when_0}{A logical}
}
\value{
List with updated var_levels and var_nlevels, and modified dt
}
\description{
handle force_generate_1_when_0
}
\keyword{internal}
