% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{handle_two_levels}
\alias{handle_two_levels}
\title{handle_two_levels}
\usage{
handle_two_levels(
  dt_wide,
  variable_interest,
  varstrat,
  patient_id,
  digits_central,
  digits_sd,
  show_metric,
  force_parametric_test,
  force_non_parametric_test,
  global_summary,
  dt
)
}
\arguments{
\item{dt_wide}{A data.table wide format}

\item{variable_interest}{A character}

\item{varstrat}{A character}

\item{patient_id}{A character}

\item{digits_central}{A integer}

\item{digits_sd}{A integer}

\item{show_metric}{A character}

\item{force_parametric_test}{A logical}

\item{force_non_parametric_test}{A logical}

\item{global_summary}{A logical}

\item{dt}{A data.table}
}
\value{
a two level description
}
\description{
handle_two_levels
}
\keyword{internal}
