% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modified_sleep}
\alias{modified_sleep}
\title{modified_sleep data}
\format{
## `modified_sleep`
A data frame with 40 rows and 6 columns:
\describe{
  \item{extra, Income, extra_with_missings, mesure1, mesure2, var1}{Numerics}
  \item{group, fact1, fact1_na, fact2, fact3, var2}{Factors}
  \item{visites_2, visites_4, visites_5}{Factors, visites factors}
  \item{ID, ID2, ID_group}{Factor, to identifiy individuals information}
  \item{all_zero_values}{Numerics, trap variables to test errors in functions}
}
}
\usage{
modified_sleep
}
\description{
Based on stats::sleep dataset, with somes modification, especially
the add of a visites variable.
}
\keyword{datasets}
