% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modified_state}
\alias{modified_state}
\title{modified_state data}
\format{
## `modified_state`
A data frame with 50 rows and ... columns:
\describe{
  \item{Population, Income, Illiteracy, Life Exp, Murder, HS Grad, Frost, Area}{Numerics, Population size, Income in USD, ... }
  \item{state.division, state.region, election}{Factors, States qualitative informations...}
  \item{binary_test}{Factor, binary information}
  \item{yes_no_french_question}{Factor, binary yes no information}
  \item{all_count_zero}{Factor, but with only one modality "0"}
  \item{var_ab_NAN, special_condition}{Factor, completely NA for one level of election (red)}
  \item{special_measures}{Numerics, completely NA for one level of election (red)}
  \item{var_conti_trap, var_quali_trap, zero_levels}{trap variables to test errors in functions}
}
}
\source{
<https://www.rdocumentation.org/packages/datasets/versions/3.6.2/topics/state>
}
\usage{
modified_state
}
\description{
Based on datasets::states* datasets, with somes modification, especially
the add of a binary test variable, add ... (see bellow)
}
\keyword{datasets}
