% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{perform_marginal_homogeneity_test}
\alias{perform_marginal_homogeneity_test}
\title{Effectuer le test d'homogénéité marginale}
\usage{
perform_marginal_homogeneity_test(dt, variable_interest, varstrat, patient_id)
}
\arguments{
\item{dt}{A data.frame}

\item{variable_interest}{A character}

\item{varstrat}{A character}

\item{patient_id}{A character}
}
\value{
Eval marginal homogeneity test
}
\description{
perform_marginal_homogeneity_test
}
\keyword{internal}
