% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{perform_paired_factorial_test}
\alias{perform_paired_factorial_test}
\title{perform_paired_factorial_test}
\usage{
perform_paired_factorial_test(
  dt,
  variable_interest,
  varstrat,
  patient_id,
  varstrat_levels
)
}
\arguments{
\item{dt}{A data.frame}

\item{variable_interest}{A character}

\item{varstrat}{A character}

\item{patient_id}{A character}

\item{varstrat_levels}{A character}
}
\value{
Choose the appropriate paired test
}
\description{
Selecting and carrying out the appropriate statistical test
}
\keyword{internal}
