% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_file_production.R
\name{quanti_sheet}
\alias{quanti_sheet}
\title{quanti_sheet}
\usage{
quanti_sheet(
  dataframe,
  vars_quanti,
  varstrat,
  precision,
  signif_digits,
  force_non_parametric_test,
  force_parametric_test,
  show_metric,
  global_summary,
  show_OR,
  show_SMD,
  detail_NB_measures,
  light_contents,
  do_test,
  verbose
)
}
\arguments{
\item{dataframe}{A data.frame. Columns must be well formatted with factor or 
numeric class (important). will be converted as data.table inside functions.}

\item{vars_quanti}{A vector of characters. Names of dataframe's columns to describe.
Only consider numerics (subset of vars).}

\item{varstrat}{A character. Default NULL. Name of the stratification variable,
making groups to compare.
Only one varstrat is expected in general.
About 2 crossed varstrats : you can provide "var1*var2" to get a description 
of the 2nd varstrat for each level of the 1st varstrat.}

\item{precision}{Precision mode: "auto" (adaptive) or numeric (fixed)
Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer, Integer indicating the number of 
decimal places (signif) for pvalues.}

\item{force_non_parametric_test}{A logical. 
You can turn it TRUE if you want to force the use of
non parametric test, whatever shapiro test said about normality. 
(so will use and show medians instead of means)}

\item{force_parametric_test}{A logical.
You can turn it TRUE if you want to force the use of
parametric test, whatever shapiro test said about normality. 
(so will use and show means instead of medians)}

\item{show_metric}{A character,
What is the metric to show in cell content ?
"auto" = mean or median, automatic choice according shapiro test,
"mean" = mean +/- sd forced, whatever shapiro said,
"median" = media [Q1;Q3] forced, whatever shapiro said.
Caution, if you force_non_parametric_test as TRUE, show_metric is forced 
as 'median' and if you force_parametric_test as TRUE, show_metric is forced 
as 'mean', to be consistent.}

\item{global_summary}{A logical. Do you want to get global summary.}

\item{show_OR}{A logical. Default FALSE. Do you want to get OR, IC and p
(when varstrat is a binary fact).}

\item{show_SMD}{A logical, Default FALSE. If turn TRUE, add SMD column
(standardized mean difference) between 2 groups.}

\item{detail_NB_measures}{A logical, Default FALSE.
If turn TRUE, N will be shown with detail one column by group levels's order.}

\item{light_contents}{A logical, Default TRUE. If FALSE, the information like 
Nb_mesures, Valeurs manquantes et p will be repeated.}

\item{do_test}{A logical, Default TRUE will apply bivariate statistical tests.
do_test = FALSE => show_SMD and show_OR trun FALSE}

\item{verbose}{A logical, Default TRUE. Show message. 
Do you want to work in silence? Turn it FALSE.}
}
\value{
list with formatted quantitative variables description
}
\description{
Compute quantitative variables sheet for Excel output
}
\details{
Internal function that processes quantitative variables and prepares
the formatted table for Excel export, with optional statistical tests.
}
