% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_file_paired_production.R
\name{quanti_sheet_paired}
\alias{quanti_sheet_paired}
\title{quanti_sheet_paired}
\usage{
quanti_sheet_paired(
  dataframe,
  vars_quanti,
  varstrat,
  patient_id,
  precision,
  signif_digits,
  global_summary,
  force_non_parametric_test,
  force_parametric_test,
  show_metric,
  detail_NB_measures,
  do_test,
  verbose
)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table will be converted into data.table.
Columns must be well formated with factor or numeric class (important).}

\item{vars_quanti}{Vector of characters. Subset of vars : vector of quantitative variable names}

\item{varstrat}{A characters. Not null. Names of the stratification variable, 
making groups to compare (*time* or *visits* for instance).
The repeated measures must be presented *in line* (for instance, "V1", "V2", "V3" 
will be set on 3 lines for each individuals, in the column "visits").
About 2 *crossed varstrats* : you can provide "var1*var2", and so
get a description of the 2nd varstrat (var2) for each level of
the 1st varstrat (var1).}

\item{patient_id}{A character.
Name of patient identifier (id) column.
The repeated measures must be presented in line 
(for instance, "id1" will be set on 3 line if he has 3 visits).}

\item{precision}{mode: "auto" (adaptive) or numeric (fixed)
Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer. Integer indicating the number of 
decimal places (signif) for p-values.}

\item{global_summary}{A logical. Default FALSE Do you want to get global summary.
Caution! global_summary on longitudinal data is not really relevant... 
but ok if you want it, you can.}

\item{force_non_parametric_test}{A logical. Default FALSE. 
You can turn it TRUE if you want to force the use of
non parametric test, whatever shapiro test said about normality.}

\item{force_parametric_test}{A logical. Default FALSE. 
You can turn it TRUE if you want to force the use of
parametric test, whatever shapiro test said about normality. 
(so will use and show means instead of medians)
This may be useful when considering the central limit theorem or small deviations.}

\item{show_metric}{A character. What is the metric to show in 
cell content ?
"auto" = mean or median, automatic choice according shapiro test,
"mean" = mean +/- sd forced, whatever shapiro said,
"median" = media [Q1;Q3] forced, whatever shapir said.
Caution, if you force_non_parametric_test as TRUE, show_metric is forced 
as 'median' and if you force_parametric_test as TRUE, show_metric is forced 
as 'mean', to be consistent.}

\item{detail_NB_measures}{A logical,
If turn TRUE, N will be shown with detail (N1 + N2 + ...) for each group.}

\item{do_test}{A logical, 
To prevent the statistical test from returning a result, set the value to FALSE.}

\item{verbose}{A logical, Default TRUE. Show message. 
Do you want to work in silence? Turn it FALSE.}
}
\value{
list with formatted quantitative variables description
}
\description{
Compute quantitative variables sheet for Excel output.
}
