% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2Excel-package.R
\docType{package}
\name{R2Excel-package}
\alias{R2Excel}
\alias{R2Excel-package}
\title{R2Excel: R2Excel : Production Of Statistical Tables in *R*, Save Into (*2*) *Excel* File}
\description{
The goal of *R2Excel* is to produce Statistical Tables in *R*, Save Into (*2*) *Excel* File. With this package, we don't propose any revolutionary statistical models or tests, but we do recommend that you use this tool for the first stages of analysis in a clinical study. Once you have your databases, you will naturally want to produce descriptive tables, and the first univariate or bivariate tests (if groups are present). We can help you to standardise these tables, while using the appropriate tests (implementation and verification of certain hypotheses, paired tests or captured messages). The main functions produce descriptive, homogeneous and harmonious statistical tables, as well as statistical tests. This package is used to generate statistical reports produced by the Methodology/Biostatistics Team of the GHICL DRCI (Lomme, France).
}
\keyword{internal}
