% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tables.R
\name{simplify_binary_table}
\alias{simplify_binary_table}
\title{simplify_binary_table}
\usage{
simplify_binary_table(vari_tab)
}
\arguments{
\item{vari_tab}{A data.frame.}
}
\value{
vari_tab simplifyed
}
\description{
simplify_binary_table
}
