% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{simplify_factorial_table}
\alias{simplify_factorial_table}
\title{simplify_factorial_table}
\usage{
simplify_factorial_table(effectif_tab)
}
\arguments{
\item{effectif_tab}{A data.frame}
}
\value{
simplified table
}
\description{
Simplify the table if required (binary variables)
}
\keyword{internal}
