% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SMD_tables.R
\name{smd_bin}
\alias{smd_bin}
\title{Standardized mean difference for binary variables}
\usage{
smd_bin(dataframe, varstrat, vars)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare.}

\item{vars}{A vector of characters. Names of dataframe's columns.}
}
\value{
A table of SMD values for each vars.
}
\description{
see more details of interpretations and use cases in `compute_SMD_table` documentation.
}
\details{
FYI : Not called in `compute_SMD_table` because `smd_bin` provides same results than `smd_cat` in binary cases.
}
\examples{
\dontrun{
smd_bin(
  modified_state,
  varstrat = "election", 
  vars = c(
    "yes_no_french_question", "binary_test"
  ) 
)
}
}
