% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SMD_tables.R
\name{smd_cat}
\alias{smd_cat}
\title{Standardized mean difference for qualitative variables}
\usage{
smd_cat(dataframe, varstrat, vars)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare.}

\item{vars}{A vector of characters. Names of dataframe's columns.}
}
\value{
A table of SMD values for each vars.
}
\description{
see more details of interpretations and use cases in `compute_SMD_table` documentation.
}
\examples{
\dontrun{
smd_cat(
  modified_state,
  varstrat = "election", 
  vars = c(
    "yes_no_french_question", "binary_test",
    "state.division", "state.region"
  ) 
)
}
}
