% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SMD_tables.R
\name{smd_num}
\alias{smd_num}
\title{Standardized mean difference for numeric (continuous) variables}
\usage{
smd_num(dataframe, varstrat, vars)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table}

\item{varstrat}{A character. Default NULL. Name of the stratification variable, making groups to compare.}

\item{vars}{A vector of characters. Names of dataframe's columns.}
}
\value{
A table of SMD values for each vars.
}
\description{
see more details of interpretations and use cases in `compute_SMD_table` documentation.
}
\examples{
\dontrun{
smd_num(
  modified_state,
  varstrat = "election", 
  vars = c(
    "Life Exp", "Frost", "Area"
  ) 
)
}
}
