% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stats.R
\name{test_means}
\alias{test_means}
\title{test means}
\usage{
test_means(
  dataframe,
  vars = setdiff(colnames(dataframe), varstrat),
  varstrat,
  force_non_parametric_test = FALSE,
  digits = 2,
  signif_digits = 4
)
}
\arguments{
\item{dataframe}{A data.frame containing data.}

\item{vars}{A vector of characters. Names of dataframe's factorial columns to describe.}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare.}

\item{force_non_parametric_test}{A logical. Default FALSE. You can turn it TRUE if you want to force the use of
non parametric test, whatever shapiro test said about normality.}

\item{digits}{A integer, Default 2. Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer, Default 4. Integer indicating the number of decimal places (signif) for pvalues.}
}
\value{
A list of 3 objects
  results = data.frame with test, P-values, explicit decision for each vars
  details = list with details, various objects.
  selection = vector of variables with p-values < 0.2
}
\description{
Test equality of means.
Deploy the tests Wilcoxon-Mann-Whitney or kruskal-Wallis for difference of means with a variable of groups for non normal data.
Otherwise Deploy the tests Student or anova.
}
\examples{
\dontrun{
test_means(
  dataframe = modified_state[, c("Population", "Income", "election")],
  vars = c("Population", "Income"),
  varstrat = "election"
)
}
}
