% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_stats.R
\name{test_proportions}
\alias{test_proportions}
\title{test proportions}
\usage{
test_proportions(
  dataframe,
  vars = setdiff(colnames(dataframe), varstrat),
  varstrat,
  digits = 2,
  signif_digits = 4
)
}
\arguments{
\item{dataframe}{A data.frame containing data.}

\item{vars}{A vector of characters. Names of dataframe's factorial columns to describe.}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare.}

\item{digits}{A integer, Default 2. Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer, Default 4. Integer indicating the number of decimal places (signif) for pvalues.}
}
\value{
A list of 3 objects
  results = data.frame with test, p-values, explicit decision for each vars
  details = list with details, various objects.
  selection = vector of variables with p-values < 0.2
}
\description{
Test contingency tables.
Deploy the tests Khi 2 and fisher for proportions with a variable of reference
}
\examples{
\dontrun{
test_proportions(
  dataframe = modified_state[, c("state.division", "state.region", "election")],
  vars = c("state.region", "state.division"),
  varstrat = "election"
)
}
}
