% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{add_global_summary}
\alias{add_global_summary}
\title{add_global_summary}
\usage{
add_global_summary(effectif_tab, dt, variable_interest, digits = 1)
}
\arguments{
\item{effectif_tab}{A data.frame}

\item{dt}{A data.frame}

\item{variable_interest}{A character}

\item{digits}{An integer}
}
\value{
formated table of frequencies
}
\description{
Ajouter la colonne Population_total si demandée
}
\keyword{internal}
