% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{add_global_summary}
\alias{add_global_summary}
\title{Ajouter la colonne Population_total si demandée}
\usage{
add_global_summary(effectif_tab, dt, variable_interest, digits = 1)
}
\description{
Ajouter la colonne Population_total si demandée
}
