% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_normality}
\alias{check_normality}
\title{Tester la normalité avec gestion d'erreurs}
\usage{
check_normality(x, return_messages = FALSE)
}
\arguments{
\item{x}{Vecteur numérique à tester}

\item{return_messages}{Logical, retourner les messages d'erreur ?}
}
\value{
Logical (normalité) ou liste avec is_normal et message
}
\description{
Tester la normalité avec gestion d'erreurs
}
\keyword{internal}
