% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tables.R
\name{compute_continuous_table}
\alias{compute_continuous_table}
\title{compute continuous table}
\usage{
compute_continuous_table(
  dataframe,
  vars = setdiff(colnames(dataframe), varstrat),
  varstrat = NULL,
  stats_choice = c("mean", "sd", "median", "Q1", "Q3", "min", "max", "N", "N_NA",
    "Nb_mesures", "is_Normal"),
  digits = 2
)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table will be converted into data.frame.}

\item{vars}{A vector of characters. Names of dataframe's continuous columns to describe.}

\item{varstrat}{A character. Default NULL. Name of the stratification variable, making groups to compare.}

\item{stats_choice}{A vector of characters. Default provide all usual statistics to describe continuous variables,
namely 'c("mean", "sd", "median", "Q1", "Q3", "min", "max", "N", "N_NA", "Nb_mesures", "is_Normal")'}

\item{digits}{A integer, Default 2. Integer indicating the number of decimal places (round).}
}
\value{
A final table or a list of tables if varstrat not null.
}
\description{
provide descriptive statistics table for continuous data
}
\examples{
\dontrun{
compute_continuous_table(
  dataframe = modified_state,
  vars = c(
    "Population", "Income", "Illiteracy", "Life Exp", "Murder",
    "HS Grad", "Frost", "Area"
  ),
  varstrat = "election",
  digits = 2
)
}
}
