% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{compute_frequency_table}
\alias{compute_frequency_table}
\title{Calculer le tableau de fréquences avec effectifs et pourcentages}
\usage{
compute_frequency_table(
  dt,
  variable_interest,
  varstrat,
  varstrat_levels,
  digits = 1
)
}
\description{
Calculer le tableau de fréquences avec effectifs et pourcentages
}
