% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{compute_paired_continuous_table_and_test}
\alias{compute_paired_continuous_table_and_test}
\title{compute continuous table for paired data}
\usage{
compute_paired_continuous_table_and_test(
  dataframe,
  variable_interest,
  varstrat,
  digits = 2,
  signif_digits = 4,
  patient_id = "patientid",
  global_summary = TRUE,
  force_non_parametric_test = FALSE,
  metric_show = "auto",
  test_more_2_levels = FALSE
)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table will be converted into data.table.}

\item{variable_interest}{A character. Name of dataframe's continuous columns to describe and test.}

\item{varstrat}{A character. Name of the stratification variable, making groups to compare (time or visites).
the repeated measures must be presented in line (for instance, "V1", "V2", "V3" will be set on 3 line for each individuals)}

\item{digits}{A integer, Default 1. Integer indicating the number of decimal places (round).}

\item{signif_digits}{A integer, Default 4. Integer indicating the number of decimal places (signif) for pvalues.}

\item{patient_id}{A character., Default "patientid". Name of identifiant patient id column.
the repeated measures must be presented in line (for instance, "id1" will be set on 3 line if he has 3 visits)}

\item{global_summary}{A logical. Default TRUE. Do you want to get global summary.
Caution! global_summary on longitudinal data is not really relevant... but ok if you want it, you can.}

\item{force_non_parametric_test}{A logical. Default FALSE. You can turn it TRUE if you want to force the use of
non parametric test, whatever shapiro test said about normality.}

\item{metric_show}{A character, Default "auto". What is the metric to show in cell content ?
"auto" = mean or median, automatic choice according statistic test,
"mean" = mean +/- sd forced, whatever shapiro said,
"median" = media [Q1;Q3] forced, whatever shapiro said.}

\item{test_more_2_levels}{A logicial, Default FALSE. Do not return stat test if varstrat has more than 2 levels}
}
\value{
a list of 2 elements :
  line_res : the data.frame with one line containing needed results
  test_result : the test object if more detail wanted about the stat test.
}
\description{
provide descriptive statistics table for continuous data with a paired level (time or visites)
}
\examples{
\dontrun{
## Paired t-test
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "extra",
  varstrat = "visites_2",
  patient_id = "ID2",
  digits = 2, signif_digits = 2,
  global_summary = FALSE
)
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "extra_with_missings",
  varstrat = "visites_2",
  patient_id = "ID2",
  digits = 2, signif_digits = 2,
  global_summary = FALSE
)
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "extra_with_missings",
  varstrat = "visites_2",
  patient_id = "ID2",
  digits = 2, signif_digits = 2,
  global_summary = TRUE
)
## Anova on paired data
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "extra",
  varstrat = "visites_4",
  patient_id = "ID",
  digits = 2, signif_digits = 2,
  global_summary = FALSE
)
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "extra",
  varstrat = "visites_5",
  patient_id = "ID",
  digits = 2, signif_digits = 2,
  global_summary = FALSE
)
## test no normal measure : Paired Wilcoxon signed-rank test
compute_paired_continuous_table_and_test(
  dataframe = modified_sleep,
  variable_interest = "mesure3",
  varstrat = "visites_2",
  patient_id = "ID2",
  digits = 2, signif_digits = 2,
  global_summary = FALSE
)
}
}
