% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tables.R
\name{compute_SMD_table}
\alias{compute_SMD_table}
\title{compute SMD}
\usage{
compute_SMD_table(dataframe, vars, varstrat, digits = 2)
}
\arguments{
\item{dataframe}{A data.frame. tibble or data.table}

\item{vars}{A vector of characters. Names of dataframe's numeric columns.}

\item{varstrat}{A character. Default NULL. Name of the stratification variable, making groups to compare.}

\item{digits}{A integer, Default 2. Integer indicating the number of decimal places (round).}
}
\value{
A table of SMD values for each vars.
}
\description{
standardized mean difference
}
\details{
Interpretation of SMD values : The generally accepted thresholds for interpreting effect size are :  
SMD < 0.1: negligible difference,  
0.1 ≤ SMD < 0.2: small difference,  
0.2 ≤ SMD < 0.5: moderate difference,  
SMD ≥ 0.5: large difference.  
These thresholds help to assess the balance between groups, especially after matching or adjustment procedures.
}
\examples{
\dontrun{
compute_SMD_table(
  dataframe = modified_state,
  vars = c(
    "Population", "Income", "Illiteracy", "Life Exp", "Murder",
    "HS Grad", "Frost", "Area"
  ),
  varstrat = "election",
  digits = 2
 )
}
}
