% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_file_production.R
\name{date_sheet}
\alias{date_sheet}
\title{date_sheet}
\usage{
date_sheet(dataframe, varstrat, vars_dates, crossed_varstrat)
}
\arguments{
\item{dataframe}{Data.frame containing the data}

\item{varstrat}{Stratification variable name (can be "", NULL, or a variable name)}

\item{vars_dates}{Character vector of dates variable names}

\item{crossed_varstrat}{Logical}
}
\value{
list with formatted dates variables description
}
\description{
Compute dates variables sheet for Excel output
}
\details{
Internal function that processes dates variables and prepares
the formatted table for Excel export
}
\keyword{internal}
