% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_cell_content}
\alias{format_cell_content}
\title{Formater le contenu de cellule selon la métrique}
\usage{
format_cell_content(stats, metric = "mean")
}
\description{
Formater le contenu de cellule selon la métrique
}
