% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_effectif_table}
\alias{format_effectif_table}
\title{Créer le tableau d'effectifs formaté}
\usage{
format_effectif_table(
  sumup,
  variable_interest,
  varstrat,
  varstrat_levels,
  keep_missing_line
)
}
\arguments{
\item{sumup}{A data.frame}

\item{variable_interest}{A character}

\item{varstrat}{A character}

\item{varstrat_levels}{A character}

\item{keep_missing_line}{A logical}
}
\value{
formated table of frequencies
}
\description{
format_effectif_table
}
\keyword{internal}
