% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{format_global_summary}
\alias{format_global_summary}
\title{Formater la description de population globale}
\usage{
format_global_summary(
  dataframe,
  variable,
  metric = "mean",
  patient_id = NULL,
  digits_central,
  digits_sd
)
}
\description{
Formater la description de population globale
}
