% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dates}
\alias{get_dates}
\title{get dates}
\usage{
get_dates(dataframe, vars = colnames(dataframe))
}
\arguments{
\item{dataframe}{A data.frame.}

\item{vars}{A vector of characters. Names of dataframe to test.}
}
\value{
A vector of dates : the names of date columns
}
\description{
provide the names of date columns
}
\examples{
\dontrun{

}
}
