% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{handle_force_generate}
\alias{handle_force_generate}
\title{Gérer force_generate_1_when_0}
\usage{
handle_force_generate(dt, variable_interest, force_generate_1_when_0)
}
\value{
Liste avec var_levels et var_nlevels mis à jour, et dt modifié
}
\description{
Gérer force_generate_1_when_0
}
