% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{perform_marginal_homogeneity_test}
\alias{perform_marginal_homogeneity_test}
\title{Effectuer le test d'homogénéité marginale}
\usage{
perform_marginal_homogeneity_test(dt, variable_interest, varstrat, patient_id)
}
\description{
Effectuer le test d'homogénéité marginale
}
