% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{perform_mcnemar_test}
\alias{perform_mcnemar_test}
\title{Effectuer le test de McNemar}
\usage{
perform_mcnemar_test(mcnemar_dt, varstrat_levels, variable_interest)
}
\description{
Effectuer le test de McNemar
}
