% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{perform_paired_factorial_test}
\alias{perform_paired_factorial_test}
\title{Choisir et effectuer le test statistique approprié}
\usage{
perform_paired_factorial_test(
  dt,
  variable_interest,
  varstrat,
  patient_id,
  varstrat_levels
)
}
\description{
Choisir et effectuer le test statistique approprié
}
