% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_paired_tables.R
\name{simplify_factorial_table}
\alias{simplify_factorial_table}
\title{Simplifier le tableau si demandé (variables binaires)}
\usage{
simplify_factorial_table(effectif_tab)
}
\description{
Simplifier le tableau si demandé (variables binaires)
}
